fixed weights v2

+++ homework3/homework/train_fcn.py
@@ -34,7 +34,7 @@ def train(args):
     if args.schedule_lr:
         scheduler = torch.optim.lr_scheduler.ReduceLROnPlateau(optimizer, 'max', patience=50)
 
-    weights = [1 / num for num in DENSE_CLASS_DISTRIBUTION]
+    weights = torch.FloatTensor([1 / num for num in DENSE_CLASS_DISTRIBUTION])
     loss = torch.nn.CrossEntropyLoss(weight=weights)
     global_step = 0
